datus-agent/pyproject.toml
# Project basic information configuration, required.
[project]
name = "datus-agent"
version = "0.1.4"
description = "Datus Agent"
requires-python = ">=3.12"
authors = [
    {name = "Datus Team"}
]

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple"
default = true
[tool.uv.pip]
index-url = "https://test.pypi.org/simple"

# uv tool run black
[tool.black]
line-length = 120
target-version = ["py312"]

[tool.flake8]
max-line-length = 120
ignore = "E203, E266, E501, W503"
max-complexity = 10
select = "B,C,E,F,W,T4,B9"
exclude = "mcp/"

[tool.pytest]
pythonpath = "."
testpaths = ["tests"]

[tool.uv]
default-groups = ["core"]

[dependency-groups]
# default dependencies
core = [
    "python-dotenv==1.0.0",
    "pandas==2.1.4",
    "sqlalchemy==2.0.23",
    "sqlglot>=26.12.0",
    "snowflake-connector-python>=3.6.0",
    "pyyaml==6.0.1",
    "langsmith>=0.0.77",
    "structlog>=23.1.0",
    "openai>=1.12.0",
    "httpx[socks]==0.27.2",
    "tantivy>=0.22.2",
    "aiohttp>=3.11.16",
    "xlsxwriter>=3.2.2",
    "tiktoken>=0.9.0",
    "openai-agents==0.0.11",
    "Markdown==3.8",
    "sqllineage>=0.20.0",
    "lancedb==0.18.0",
    "pylance==0.22.0",
    "datasets>=3.5.1",
    "transformers>=4.51.3",
    "sentence-transformers==4.1.0",
    # this verison is required by pandas 2.1.4
    "pyarrow<19.0.0",
    # This library is required by lancedb
    "rich==14.0.0",
    "prompt_toolkit>=3.0.51",
    "textual==3.2.0",
    "anthropic==0.51.0",
    "duckdb-engine>=0.17.0",
    "snowflake-sqlalchemy>=1.7.3",
    # Dependencies by snowflake-sqlalchemy
    "opentelemetry-api>=1.33.1",
    # MCP (Model Context Protocol) support
    "mcp>=1.9.2",
    "anyio>=4.9.0",
    # sentence-transformers and lancedb depends on the torch. If your mac is not an arm chip, then version 2.2.2 should be used under python 3.12.
    # otherwise, you can use the latest version of torch(2.7.0).
    "torch==2.2.2",
    "mysqlclient==2.2.7",
    "pymysql>=1.1.1",
]
# uv sync --dev
dev = [
    "pytest==8.0.2",
    "pytest-asyncio>=0.23.8",
    "black==23.12.1",
    "mypy==1.8.0",
    "tqdm>=4.27.0",
    "pytest-xdist>=3.7.0",
    "pre-commit>=4.2.0",
]
